+++ spec/policies/facility_group_policy_spec.rb
@@ -12,10 +12,12 @@ RSpec.describe FacilityGroupPolicy do
   let(:supervisor) { FactoryBot.create(:admin, :supervisor, admin_access_controls: [AdminAccessControl.new(access_controllable: facility_group_in_organization)]) }
   let(:analyst) { FactoryBot.create(:admin, :analyst) }
 
-  permissions :index? do
+  permissions :index?, :new?, :create? do
     it "permits owners and organization owners" do
+      new_facility_group = organization.facility_groups.new
       expect(subject).to permit(owner, FacilityGroup)
       expect(subject).to permit(organization_owner, FacilityGroup)
+      expect(subject).to permit(organization_owner, new_facility_group)
     end
 
     it "denies supervisors and analysts" do
@@ -24,7 +26,7 @@ RSpec.describe FacilityGroupPolicy do
     end
   end
 
-  permissions :show?, :new?, :create?, :update?, :edit?, :destroy? do
+  permissions :show?, :update?, :edit?, :destroy? do
     it "permits owners for all facility groups" do
       expect(subject).to permit(owner, facility_group_in_organization)
       expect(subject).to permit(owner, facility_group_outside_organization)