+++ CHANGELOG.md
@@ -4,11 +4,30 @@
 ### Fixed
 ### Deprecated 
 ### Removed 
+### Security
+
+# 2019-01-15-2
+## Portal
+### Added
+### Changed
+### Fixed
+- Update controller allow creating new facility groups
+### Deprecated
+### Removed
 ### Security 
 
+# 2019-01-15-1
+## Portal
+### Added
+### Changed
+### Fixed
+- Fix organization owners policy to allow creating new records in their organizations
+### Deprecated
+### Removed
+### Security
 
 # 2019-01-14-1
-
+## Portal
 ### Added
 ### Changed
 - Move Invite Owner button before Invite Organization Owner on Admins page